# pkg/modules/pipes/implpipes/handler.go
package implpipes

import (
	"encoding/json"
	"net/http"

	"github.com/gear6io/pragmata/pkg/http/render"
	"github.com/gear6io/pragmata/pkg/modules/pipes"
	"github.com/gear6io/pragmata/pkg/types/pipetypes"
)

type handler struct {
	module pipes.Module
}

// NewHandler wraps a Module as an HTTP handler.
func NewHandler(mod pipes.Module) pipes.Handler {
	return &handler{module: mod}
}

func (h *handler) CreatePipe(w http.ResponseWriter, r *http.Request) {
	var pipe pipetypes.Pipe
	if !decodeJSON(w, r, &pipe) {
		return
	}
	created, err := h.module.CreatePipe(r.Context(), &pipe)
	if err != nil {
		render.Error(w, http.StatusInternalServerError, err.Error())
		return
	}
	render.Success(w, http.StatusCreated, created)
}

func (h *handler) ListPipes(w http.ResponseWriter, r *http.Request) {
	list, err := h.module.ListPipes(r.Context())
	if err != nil {
		render.Error(w, http.StatusInternalServerError, err.Error())
		return
	}
	if list == nil {
		list = []*pipetypes.Pipe{}
	}
	render.Success(w, http.StatusOK, list)
}

func (h *handler) GetPipe(w http.ResponseWriter, r *http.Request) {
	name := pipes.PipeName(r)
	pipe, err := h.module.GetPipe(r.Context(), name)
	if err != nil {
		render.Error(w, http.StatusNotFound, err.Error())
		return
	}
	render.Success(w, http.StatusOK, pipe)
}

func (h *handler) UpdatePipe(w http.ResponseWriter, r *http.Request) {
	name := pipes.PipeName(r)
	var pipe pipetypes.Pipe
	if !decodeJSON(w, r, &pipe) {
		return
	}
	updated, err := h.module.UpdatePipe(r.Context(), name, &pipe)
	if err != nil {
		render.Error(w, http.StatusInternalServerError, err.Error())
		return
	}
	render.Success(w, http.StatusOK, updated)
}

func (h *handler) DeletePipe(w http.ResponseWriter, r *http.Request) {
	name := pipes.PipeName(r)
	if err := h.module.DeletePipe(r.Context(), name); err != nil {
		render.Error(w, http.StatusInternalServerError, err.Error())
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func decodeJSON(w http.ResponseWriter, r *http.Request, v any) bool {
	if err := json.NewDecoder(r.Body).Decode(v); err != nil {
		render.Error(w, http.StatusBadRequest, "invalid JSON: "+err.Error())
		return false
	}
	return true
}